Successfully tested ElectricalConfiguration to View Binding
Use ERV View Model to fully test if the visibility binding were working as intended. They WERE! And we can use the ERV View Model as an example.

Also created ElectricalPanelViewModelFactory. This is used for passing ElectricalConfiguration to ElectricalViewModel from within the main EquipmentViewModel. Use this as design for all other panel ViewModels that use configuration.